requirements.txt
Django==3.2
Pillow==8.2.0
mysqlclient==2.0.3
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22
requests==2.25.1
djangorestframework==3.12.4
django-oauth-toolkit==1.5.0
isort==5.9.3
black==21.8b0
django-environ==0.4.5
django-oauth-toolkit django-cors-headers